+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Changelog
 
+## v0.0.3
+### Fixed
+- Remove base64 encoding on write in the KinesisClient.
+- Add UTF-8 encoding for StringIO object in NYPLAvro.
+
 ## v0.0.0
 
 Initial version. Includes `kinesis_client`, `kms_client`, `nypl_avro`, `platform_api_client`,